update: test recipe
