+++ tests/utils.py
@@ -7,6 +7,7 @@ from torch.nn.functional import relu
 
 from pytorch_optimizer.base.type import LOSS
 from pytorch_optimizer.optimizer import AdamW, Lookahead, OrthoGrad, ScheduleFreeWrapper
+from pytorch_optimizer.optimizer.alig import l2_projection
 
 
 class LogisticRegression(nn.Module):
@@ -130,3 +131,26 @@ def sphere_loss(x: torch.Tensor) -> torch.Tensor:
 def build_model(use_complex: bool = False):
     torch.manual_seed(42)
     return ComplexLogisticRegression() if use_complex else LogisticRegression(), nn.BCEWithLogitsLoss()
+
+
+def build_optimizer_parameter(parameters, optimizer_name, config):
+    if optimizer_name == 'AliG':
+        config.update({'projection_fn': lambda: l2_projection(parameters, max_norm=1)})
+    elif optimizer_name in ('Muon', 'AdaMuon'):
+        hidden_weights = [p for p in parameters if p.ndim >= 2]
+        hidden_gains_biases = [p for p in parameters if p.ndim < 2]
+
+        parameters = [
+            {'params': hidden_weights, 'use_muon': True},
+            {'params': hidden_gains_biases, 'use_muon': False},
+        ]
+    elif optimizer_name == 'AdamWSN':
+        sn_params = [p for p in parameters if p.ndim == 2]
+        regular_params = [p for p in parameters if p.ndim != 2]
+        parameters = [{'params': sn_params, 'sn': True}, {'params': regular_params, 'sn': False}]
+    elif optimizer_name == 'AdamC':
+        norm_params = [p for i, p in enumerate(parameters) if i == 1]
+        regular_params = [p for i, p in enumerate(parameters) if i != 1]
+        parameters = [{'params': norm_params, 'normalized': True}, {'params': regular_params}]
+
+    return parameters, config
